refactor general reports
