+++ app/views/reports/general_reports/_report_controls.html.haml
