+++ app/views/reports/general_reports/_report_rows.html.haml
