+++ app/views/reports/general_reports/_report_tabs.html.haml
