+++ config/routes.rb
@@ -281,7 +281,8 @@ Nucore::Application.routes.draw do
       end
     end
 
-    get "general_reports/:report_by", to: "general_reports#report_by", as: "general_reports"
+    get "general_reports/raw", to: "reports/export_raw_reports#export_all", as: "export_raw_reports"
+    get "general_reports/:report_by", to: "reports/general_reports#index", as: "general_reports"
     get "instrument_reports/:report_by", to: "instrument_reports#report_by", as: "instrument_reports"
   end
